dummyapi/requirements_mqtt.txt
flask==2.2.3
flask-cors==3.0.10
paho-mqtt==2.2.1
